raspberry/requirements.txt
Adafruit-DHT==1.4.0
gcloud==0.17.0
googleapis-common-protos==1.5.6
httplib2==0.12.0
jws==0.1.3
oauth2client==3.0.0
pkg-resources==0.0.0
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycryptodome==3.4.3
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
RPi.GPIO==0.6.5
rsa==4.0
six==1.12.0
